+++ packages/abi/_version.d.ts
@@ -1,1 +1,1 @@
-export declare const version = "5.0.0-beta.132";
+export declare const version = "5.0.0-beta.133";