+++ packages/properties/index.d.ts
@@ -1,4 +1,5 @@
 export declare function defineReadOnly(object: any, name: string, value: any): void;
+export declare function getStatic<T>(ctor: any, key: string): T;
 export declare function resolveProperties(object: any): Promise<any>;
 export declare function checkProperties(object: any, properties: {
     [name: string]: boolean;